Add more package types
