+++ src/main/java/ThatOneGuy.java
@@ -1,6 +1,9 @@
 import java.util.*;
+
+import parser.Parser;
 import tasks.*;
 import storage.Storage;
+import ui.Ui;
 
 public class ThatOneGuy {
     private TaskManager tm;
